Create temporary files in the same directory
Previously, we were creating temp files in the current working
directory, but there's no guarantee that this would be on the same
filesystem as the destination files. Since os.Rename() fails when
renaming across devices, we need to ensure that the temporary files
are on the same device. The only guaranteed place they can go is in
the same directory as the output file.