Initialize OpenSSL with OpenSSL <1.1.0
This is needed for `EVP_get_digestbyname` to work on OpenSSL<1.1.0, because we
need to manually initialize the name->digest lookup table.
This didn't get caught before probably because the default used `EVP_sha256`,
which does an implicit initialization for just sha256.

Without this, on OpenSSL 1.0, you get `Unsupported hashing algorithm.Invalid argument.`.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>