Don't default to generating DH params file
When this feature was added initially, it generated the file
unconditionally. We later added a --no-dhparams-file argument for those
consumers who explicitly didn't want it, but to avoid compatibility
issues we defaulted to continuing to produce it. This was never the
desired situation, so it will be removed now as part of SSCG 4.0.

Fixes: https://github.com/sgallagher/sscg/issues/91

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>