Avoid TOCTOU when opening files.
Instead of normalizing the paths and then checking to see if it matches
a file we already know about, sscg will now just preemprively attempt to
open the files and then perform a comparison of the st_dev and st_inode
values to determine if they are the same. This avoids a possible
Time-of-Check/Time-of-Use vulnerability where the contents of the file
changes between when it is inspected and when it is opened.

TESTS: fix incorrect return code checks

Prior to the removal of path_utils and the reworking of the output
files to avoid a time-of-check/time-of-use period, the old code was
returning exit code "1" instead of the appropriate errno values.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>