Handle opening the same file
We introduced a regression in 4.0.1 when closing the TOCTOU issue. We
incorrectly refuse to open the same file twice to save both a CA and
cert into the same location.

This changes the opening logic to always open for  both reading  and
writing to check if it is zero-length.

Also switches the EC curve tests to output the CA and service cert to a
common path to ensure we are testing this case properly.

Updates some incorrect expected behaviors in the test_dhparams_creation
test as well.

Fixes: https://github.com/sgallagher/sscg/issues/99

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>