Better handling of file opening
Don't truncate destination files until just prior to writing.

Rework the file opening logic to only care whether the file was
pre-existing and not to care if it was previously empty. Additionally,
don't truncate until we're just about to write to the files, in case
something goes wrong with the key generation or signing.

Also, if the process fails, SSCG will now delete any files it was
holding open, rather than leaving them on the filesystem.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>